+++ apps/project_test/lib/models/test.dart
@@ -1,0 +1,12 @@
+
+
+// part 'test.freezed.dart';
+// part 'test.g.dart';
+
+// @freezed
+// class Test with _$Test {
+
+//   factory Test() = _Test;
+
+//   factory Test.fromJson(Map<String, dynamic> json) => _$TestFromJson(json);
+// }