+++ packages/sup_gen_models/lib/table/extension.dart
@@ -1,7 +1,0 @@
-import 'package:sup_gen_model/table/table_model.dart';
-
-extension on TableModel {
-
-
-
-}