+++ packages/sup_gen_runner/lib/src/helpers/env.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:sup_gen_model/database_option.dart';
 
+
 DatabaseOption loadDbOptionFromEnvFile({required File envFile}) {
   if (!envFile.existsSync()) {
     throw Exception('The .env file does not exist');
@@ -19,14 +20,13 @@ DatabaseOption loadDbOptionFromEnvFile({required File envFile}) {
     final value = parts[1].trim();
     envMap[key] = value;
   }
-  // stdout.writeln('[SupGen] Loaded .env file');
-
   final keys = {
     "SUPABASE_DB_USER",
     "SUPABASE_DB_HOST",
     "SUPABASE_DB_PORT",
     "SUPABASE_DB_PASSWORD",
-    "SUPABASE_DB_SCHEMA"
+    "SUPABASE_DB_SCHEMA",
+    "SUPABASE_DB"
   };
   // check if the required keys are present
 
@@ -42,6 +42,7 @@ DatabaseOption loadDbOptionFromEnvFile({required File envFile}) {
     password: envMap['SUPABASE_DB_PASSWORD']!,
     host: envMap['SUPABASE_DB_HOST']!,
     schema: envMap['SUPABASE_DB_SCHEMA'] ?? 'public',
+    db: envMap['SUPABASE_DB'] ?? 'postgres',
     port: int.parse(envMap['SUPABASE_DB_PORT']!),
   );
 }