+++ packages/sup_gen_sqler/test/sqler_test.dart
@@ -1,16 +1,15 @@
-import 'package:sup_gen_sqler/sqler.dart';
 import 'package:test/test.dart';
 
 void main() {
   group('A group of tests', () {
-    final awesome = Awesome();
+    // final awesome = Awesome();
 
-    setUp(() {
-      // Additional setup goes here.
-    });
+    // setUp(() {
+    //   // Additional setup goes here.
+    // });
 
-    test('First Test', () {
-      expect(awesome.isAwesome, isTrue);
-    });
+    // test('First Test', () {
+    //   expect(awesome.isAwesome, isTrue);
+    // });
   });
 }