+++ packages/sup_gen_runner/CHANGELOG.md
@@ -1,3 +1,35 @@
+
+
+## 1.2.1
+
+ - Fix: update Readme
+
+## 1.2.0
+
+- Update supgen_models utility to 1.5.0
+- Enhance enums types generation
+- Handle array type generation. Ex: _text => List<String>
+- Add reverse retrieval of Enum for enums containing hyphens (-)
+
+
+## 1.0.2
+
+- Add sql function to main project
+- Remove sqler package
+- Add async gen
+- Fix builder not exiting when done
+- add fromJson to generated classes
+- Fix table generator
+- update Table Model
+- Updgrate model package
+- Improve error handling in async functions
+- Refactor codebase for better readability
+- Update dependencies to latest versions
+- Optimize performance of table generator
+- Fix minor bugs in model package
+
+
 ## 1.0.0
 
 - Initial version.
+