+++ packages/sup_gen_sqler/CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 1.0.0
 
 - Initial version.
+
+## 1.0.1
+
+- add get view definition from database
+
+## 1.0.2
+
+- updgrade models
+